Really hacky and mostly broken initial version.
Data direction, power direction and connector orientation are hardcoded.
That means, flip the USB-C around, and it won't work.
Also, the USB VBUS regulator isn't currently turning on, so VBUS is only ~3V
(whatever the battery provides). Almost nothing works at this voltage,
but the built-in Ethernet adapter of the Dell DA200 dongle does, providing
a working network!

Only tested on build 17134 for now.

WARNING: Do NOT plug in a charger or connect to your computer with this loaded,
as it will backfeed ~3V into whatever port it's plugged in, and can damage the Lumia
or the computer!
it's plugged into!