Architectural improvements and charging support
Charger capability detection is not implemented yet, but Windows
now recognizes when a charger is plugged in. A default charging
current of 500mA is used for safety, which can be overridden by setting
HKLM\SYSTEM\usbc\ChargeCurrent to the desired milliamps rating.
(Take care to set the correct *decimal* value - regedit defaults to
hexadecimal, so entering e.g. 2000 without switching to decimal
will set the current to 8192mA, instead of 2000mA!)