Regenerate the project files
I originally started working on this project in Visual Studio 2015, compiling in the Enterprise WDK command prompt. Somehow, something got messed up
in the project configuration, causing WPP tracing to be unusable with the old project files.

To fix this, I created a new KMDF driver project with the same name from the default template, and simply copied over the code from the old project.